chore: release v0.5.0-alpha.0
**!!! Please note, this currently a Alpha-Version, because many things are changed and needed to be tested!!!**
The core of the adapter keeps running on the same 0.4.1 version like before, but the configuration page has many improvements. See the list below!

Note: save content of the state 'configuration' inside the instance, as text on your disk to be able, to restore it, it case needed!

After downloading the BETA Version of the adapter, please manually proceed with uploading the adapter (this has to be done, after adapters are installed via Github)
described here: https://www.iobroker.net/#de/documentation/tutorial/adapter.md?#uploadvonadapterdateien

Have a look on Github, to verify, whats new in this Alpha release!